Rename configureNamedContext to createNamedContext.
